Move RosBag2RecordTestFixture insight test_record_params.cpp
